abstracted task functions to helper so it can be reused in the dashboard page
